for test
PiperOrigin-RevId: 263576584

+++ tfx/tools/cli/handler/airflow_handler.py
@@ -35,40 +35,50 @@ class AirflowHandler(base_handler.BaseHandler):
 
   def __init__(self, flags_dict):
     super(AirflowHandler, self).__init__(flags_dict)
-    self._handler_home_dir = os.path.join(self._handler_home_dir, 'dags', '')
+    self._handler_home_dir = self._get_handler_home('airflow')
 
-  def create_pipeline(self, overwrite: bool = False) -> None:
-    """Creates pipeline in Airflow.
+  # TODO(b/132286477): Update comments after updating methods.
+  def create_pipeline(self, overwrite: bool = False):
+    """Creates pipeline in Airflow."""
 
-    Args:
-      overwrite: Set as True to update pipeline.
-    """
     # Compile pipeline to check if pipeline_args are extracted successfully.
     pipeline_args = self.compile_pipeline()
 
-    pipeline_name = pipeline_args[labels.PIPELINE_NAME]
+    # Path to pipeline folder in airflow.
+    handler_pipeline_path = self._get_handler_pipeline_path(
+        pipeline_args[labels.PIPELINE_NAME])
 
-    self._check_pipeline_existence(pipeline_name, required=overwrite)
+    if overwrite:
+      # For update, check if pipeline exists.
+      if not tf.io.gfile.exists(handler_pipeline_path):
+        sys.exit('Pipeline "{}" does not exist.'.format(
+            pipeline_args[labels.PIPELINE_NAME]))
+    else:
+      # For create, verify that pipeline does not exist.
+      if tf.io.gfile.exists(handler_pipeline_path):
+        sys.exit('Pipeline "{}" already exists.'.format(
+            pipeline_args[labels.PIPELINE_NAME]))
 
     self._save_pipeline(pipeline_args)
 
     if overwrite:
-      click.echo('Pipeline "{}" updated successfully.'.format(pipeline_name))
+      click.echo('Pipeline "{}" updated successfully.'.format(
+          pipeline_args[labels.PIPELINE_NAME]))
     else:
-      click.echo('Pipeline "{}" created successfully.'.format(pipeline_name))
+      click.echo('Pipeline "{}" created successfully.'.format(
+          pipeline_args[labels.PIPELINE_NAME]))
 
-  def update_pipeline(self) -> None:
-    """Updates pipeline in Airflow."""
-    # Set overwrite as True to update the pipeline.
+  def update_pipeline(self):
+    # Set overwrite to true for update to make sure pipeline exists.
     self.create_pipeline(overwrite=True)
 
   def list_pipelines(self) -> None:
     """List all the pipelines in the environment."""
-    if not tf.io.gfile.exists(self._handler_home_dir):
+    dags_folder = os.path.join(self._handler_home_dir, 'dags', '')
+    if not tf.io.gfile.exists(dags_folder):
       click.echo('No pipelines to display.')
       return
-
-    pipelines_list = tf.io.gfile.listdir(self._handler_home_dir)
+    pipelines_list = tf.io.gfile.listdir(dags_folder)
 
     # Print every pipeline name in a new line.
     click.echo('-' * 30)
@@ -77,25 +87,22 @@ class AirflowHandler(base_handler.BaseHandler):
 
   def delete_pipeline(self) -> None:
     """Delete pipeline in Airflow."""
-    pipeline_name = self.flags_dict[labels.PIPELINE_NAME]
-
-    # Path to pipeline folder.
-    handler_pipeline_path = os.path.join(self._handler_home_dir, pipeline_name,
-                                         '')
+    # Path to pipeline folder in airflow.
+    handler_pipeline_path = self._get_handler_pipeline_path(
+        self.flags_dict[labels.PIPELINE_NAME])
 
     # Check if pipeline exists.
-    self._check_pipeline_existence(pipeline_name)
+    if not tf.io.gfile.exists(handler_pipeline_path):
+      sys.exit('Pipeline "{}" does not exist.'.format(
+          self.flags_dict[labels.PIPELINE_NAME]))
 
     # Delete pipeline folder.
     io_utils.delete_dir(handler_pipeline_path)
-    click.echo('Pipeline "{}" deleted successfully.'.format(pipeline_name))
+    click.echo('Pipeline "{}" deleted successfully.'.format(
+        self.flags_dict[labels.PIPELINE_NAME]))
 
   def compile_pipeline(self) -> Dict[Text, Any]:
-    """Compiles pipeline in Airflow.
-
-    Returns:
-      A python dictionary with pipeline details extracted from DSL.
-    """
+    """Compiles pipeline in Airflow."""
     self._check_pipeline_dsl_path()
     self._check_dsl_runner()
     pipeline_args = self._extract_pipeline_args()
@@ -106,18 +113,23 @@ class AirflowHandler(base_handler.BaseHandler):
 
   def create_run(self) -> None:
     """Trigger DAG in Airflow."""
-    pipeline_name = self.flags_dict[labels.PIPELINE_NAME]
-
     # Check if pipeline exists.
-    self._check_pipeline_existence(pipeline_name)
+    handler_pipeline_path = self._get_handler_pipeline_path(
+        self.flags_dict[labels.PIPELINE_NAME])
+    if not tf.io.gfile.exists(handler_pipeline_path):
+      sys.exit('Pipeline "{}" does not exist.'.format(
+          self.flags_dict[labels.PIPELINE_NAME]))
 
     # Unpause DAG.
-    self._subprocess_call(['airflow', 'unpause', pipeline_name])
+    self._subprocess_call(
+        ['airflow', 'unpause', self.flags_dict[labels.PIPELINE_NAME]])
 
     # Trigger DAG.
-    self._subprocess_call(['airflow', 'trigger_dag', pipeline_name])
+    self._subprocess_call(
+        ['airflow', 'trigger_dag', self.flags_dict[labels.PIPELINE_NAME]])
 
-    click.echo('Run created for pipeline: ' + pipeline_name)
+    click.echo('Run created for pipeline: ' +
+               self.flags_dict[labels.PIPELINE_NAME])
 
   def delete_run(self) -> None:
     """Deletes a run in Airflow."""
@@ -130,20 +142,26 @@ class AirflowHandler(base_handler.BaseHandler):
   def list_runs(self) -> None:
     """Lists all runs of a pipeline in Airflow."""
     # Check if pipeline exists.
-    pipeline_name = self.flags_dict[labels.PIPELINE_NAME]
-    self._check_pipeline_existence(pipeline_name)
-    self._subprocess_call(['airflow', 'list_dag_runs', pipeline_name])
+    handler_pipeline_path = self._get_handler_pipeline_path(
+        self.flags_dict[labels.PIPELINE_NAME])
+    if not tf.io.gfile.exists(handler_pipeline_path):
+      sys.exit('Pipeline "{}" does not exist.'.format(
+          self.flags_dict[labels.PIPELINE_NAME]))
+    self._subprocess_call(
+        ['airflow', 'list_dag_runs', self.flags_dict[labels.PIPELINE_NAME]])
 
   def get_run(self) -> None:
     """Checks run status in Airflow."""
-    pipeline_name = self.flags_dict[labels.PIPELINE_NAME]
-
     # Check if pipeline exists.
-    self._check_pipeline_existence(pipeline_name)
-
-    # Get status of all DAG runs.
+    handler_pipeline_path = self._get_handler_pipeline_path(
+        self.flags_dict[labels.PIPELINE_NAME])
+    if not tf.io.gfile.exists(handler_pipeline_path):
+      sys.exit('Pipeline "{}" does not exist.'.format(
+          self.flags_dict[labels.PIPELINE_NAME]))
     dag_runs_list = str(
-        subprocess.check_output(['airflow', 'list_dag_runs', pipeline_name]))
+        subprocess.check_output(
+            ['airflow', 'list_dag_runs',
+             self.flags_dict[labels.PIPELINE_NAME]]))
     lines = dag_runs_list.split('\\n')
     for line in lines:
       # The tokens are id, run_id, state, execution_date, state_date
@@ -151,18 +169,13 @@ class AirflowHandler(base_handler.BaseHandler):
       if self.flags_dict[labels.RUN_ID] in line:
         click.echo('run_id :' + tokens[1])
         click.echo('state :' + tokens[2])
-        break
 
   def _save_pipeline(self, pipeline_args: Dict[Text, Any]) -> None:
-    """Creates/updates pipeline folder in the handler directory.
+    """Creates/updates pipeline folder in the handler directory."""
 
-    Args:
-      pipeline_args: Pipeline details obtained from DSL.
-    """
-    # Path to pipeline folder in Airflow.
-    handler_pipeline_path = os.path.join(self._handler_home_dir,
-                                         pipeline_args[labels.PIPELINE_NAME],
-                                         '')
+    # Path to pipeline folder in airflow.
+    handler_pipeline_path = self._get_handler_pipeline_path(
+        pipeline_args[labels.PIPELINE_NAME])
 
     # If updating pipeline, first delete pipeline directory.
     if tf.io.gfile.exists(handler_pipeline_path):
@@ -175,8 +188,22 @@ class AirflowHandler(base_handler.BaseHandler):
       json.dump(pipeline_args, f)
 
     # Copy dsl to pipeline folder
-    pipeline_dsl_path = self.flags_dict[labels.PIPELINE_DSL_PATH]
     io_utils.copy_file(
-        pipeline_dsl_path,
-        os.path.join(handler_pipeline_path,
-                     os.path.basename(pipeline_dsl_path)))
+        self.flags_dict[labels.PIPELINE_DSL_PATH],
+        os.path.join(
+            handler_pipeline_path,
+            os.path.basename(self.flags_dict[labels.PIPELINE_DSL_PATH])
+            )
+        )
+
+  def _get_handler_pipeline_path(self, pipeline_name: Text) -> Text:
+    """Path to pipeline folder in airflow.
+
+    Args:
+      pipeline_name: name of the pipeline
+
+    Returns:
+      Path to pipeline folder in airflow.
+    """
+    # Path to pipeline folder in airflow.
+    return os.path.join(self._handler_home_dir, 'dags', pipeline_name)
