+++ tfx/tools/cli/handler/beam_handler.py
@@ -32,30 +32,42 @@ from tfx.utils import io_utils
 class BeamHandler(base_handler.BaseHandler):
   """Helper methods for Beam Handler."""
 
-  def create_pipeline(self, overwrite: bool = False) -> None:
-    """Creates pipeline in Beam.
+  def __init__(self, flags_dict: Dict[Text, Any]):
+    super(BeamHandler, self).__init__(flags_dict)
+    self._handler_home_dir = self._get_handler_home('beam')
 
-    Args:
-      overwrite: set as true to update pipeline.
-    """
+  def create_pipeline(self, overwrite: bool = False):
+    """Creates pipeline in Beam."""
+
     # Compile pipeline to check if pipeline_args are extracted successfully.
     pipeline_args = self.compile_pipeline()
 
-    pipeline_name = pipeline_args[labels.PIPELINE_NAME]
+    # Path to pipeline folder in beam.
+    handler_pipeline_path = self._get_handler_pipeline_path(
+        pipeline_args[labels.PIPELINE_NAME])
 
-    self._check_pipeline_existence(pipeline_name, required=overwrite)
+    if overwrite:
+      # For update, check if pipeline exists.
+      if not tf.io.gfile.exists(handler_pipeline_path):
+        sys.exit('Pipeline "{}" does not exist.'.format(
+            pipeline_args[labels.PIPELINE_NAME]))
+    else:
+      # For create, verify that pipeline does not exist.
+      if tf.io.gfile.exists(handler_pipeline_path):
+        sys.exit('Pipeline "{}" already exists.'.format(
+            pipeline_args[labels.PIPELINE_NAME]))
 
     self._save_pipeline(pipeline_args)
 
     if overwrite:
-      click.echo('Pipeline "{}" updated successfully.'.format(pipeline_name))
-    else:
-      click.echo('Pipeline "{}" created successfully.'.format(pipeline_name))
+      click.echo('Pipeline "{}" updated successfully.'.format(
+          pipeline_args[labels.PIPELINE_NAME]))
+    click.echo('Pipeline "{}" created successfully.'.format(
+        pipeline_args[labels.PIPELINE_NAME]))
 
   def update_pipeline(self) -> None:
     """Updates pipeline in Beam."""
-
-    # Set overwrite as True to update the pipeline.
+    # Set overwrite to true for update to make sure pipeline exists.
     self.create_pipeline(overwrite=True)
 
   def list_pipelines(self) -> None:
@@ -72,25 +84,22 @@ class BeamHandler(base_handler.BaseHandler):
 
   def delete_pipeline(self) -> None:
     """Deletes pipeline in Beam."""
-    pipeline_name = self.flags_dict[labels.PIPELINE_NAME]
-
     # Path to pipeline folder.
-    handler_pipeline_path = os.path.join(self._handler_home_dir, pipeline_name,
-                                         '')
+    handler_pipeline_path = self._get_handler_pipeline_path(
+        self.flags_dict[labels.PIPELINE_NAME])
 
     # Check if pipeline exists.
-    self._check_pipeline_existence(pipeline_name)
+    if not tf.io.gfile.exists(handler_pipeline_path):
+      sys.exit('Pipeline "{}" does not exist.'.format(
+          self.flags_dict[labels.PIPELINE_NAME]))
 
     # Delete pipeline folder.
     io_utils.delete_dir(handler_pipeline_path)
-    click.echo('Pipeline "{}" deleted successfully.'.format(pipeline_name))
+    click.echo('Pipeline "{}" deleted successfully.'.format(
+        self.flags_dict[labels.PIPELINE_NAME]))
 
   def compile_pipeline(self) -> Dict[Text, Any]:
-    """Compiles pipeline in Beam.
-
-    Returns:
-      pipeline_args: python dictionary with pipeline details extracted from DSL.
-    """
+    """Compiles pipeline in Beam."""
     self._check_pipeline_dsl_path()
     self._check_dsl_runner()
     pipeline_args = self._extract_pipeline_args()
@@ -101,13 +110,19 @@ class BeamHandler(base_handler.BaseHandler):
 
   def create_run(self) -> None:
     """Runs a pipeline in Beam."""
-    pipeline_name = self.flags_dict[labels.PIPELINE_NAME]
+
+    # Path to pipeline folder.
+    handler_pipeline_path = self._get_handler_pipeline_path(
+        self.flags_dict[labels.PIPELINE_NAME])
 
     # Check if pipeline exists.
-    self._check_pipeline_existence(pipeline_name)
+    if not tf.io.gfile.exists(handler_pipeline_path):
+      sys.exit('Pipeline "{}" does not exist.'.format(
+          self.flags_dict[labels.PIPELINE_NAME]))
 
     # Get dsl path from pipeline args.
-    pipeline_args_path = os.path.join(self._handler_home_dir, pipeline_name,
+    pipeline_args_path = os.path.join(self._handler_home_dir,
+                                      self.flags_dict[labels.PIPELINE_NAME],
                                       'pipeline_args.json')
     with open(pipeline_args_path, 'r') as f:
       pipeline_args = json.load(f)
@@ -134,14 +149,14 @@ class BeamHandler(base_handler.BaseHandler):
 
   def _save_pipeline(self, pipeline_args: Dict[Text, Any]) -> None:
     """Creates/updates pipeline folder in the handler directory."""
+
     # Add pipeline dsl path to pipeline args.
     pipeline_args[labels.PIPELINE_DSL_PATH] = self.flags_dict[
         labels.PIPELINE_DSL_PATH]
 
     # Path to pipeline folder in beam.
-    handler_pipeline_path = os.path.join(self._handler_home_dir,
-                                         pipeline_args[labels.PIPELINE_NAME],
-                                         '')
+    handler_pipeline_path = self._get_handler_pipeline_path(
+        pipeline_args[labels.PIPELINE_NAME])
 
     # If updating pipeline, first delete pipeline directory.
     if tf.io.gfile.exists(handler_pipeline_path):
@@ -152,3 +167,15 @@ class BeamHandler(base_handler.BaseHandler):
     with open(os.path.join(handler_pipeline_path, 'pipeline_args.json'),
               'w') as f:
       json.dump(pipeline_args, f)
+
+  def _get_handler_pipeline_path(self, pipeline_name: Text) -> Text:
+    """Path to pipeline folder in beam.
+
+    Args:
+      pipeline_name: name of the pipeline
+
+    Returns:
+      Path to pipeline folder in beam.
+    """
+    # Path to pipeline folder in beam.
+    return os.path.join(self._handler_home_dir, pipeline_name)
